# tequilapi/docs/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "The purpose of this documentation is to provide developers an insight of how to\ninteract with Mysterium Node via Tequila API.\nThis should demonstrate all the possible API calls with described parameters and responses.",
    "title": "Tequila API",
    "version": "dev"
  },
  "host": "127.0.0.1:4050",
  "paths": {
...
    "/node/latest-release": {
      "get": {
        "description": "Checks for latest Node release package and retrieves its information",
        "tags": [
          "node"
        ],
        "summary": "Latest Node release information",
        "operationId": "GetLatestRelease",
        "responses": {
          "200": {
            "description": "Latest Node release information",
            "schema": {
              "$ref": "#/definitions/LatestReleaseResponse"
            }
          },
          "500": {
            "description": "Failed to retrieve latest Node release information",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          }
        }
      }
    },
...
    "/node/provider/activity-stats": {
      "get": {
        "description": "Node activity stats",
        "tags": [
          "provider"
        ],
        "summary": "Provides Node activity stats",
        "operationId": "GetProviderActivityStats",
        "responses": {
          "200": {
            "description": "Provider activity stats",
            "schema": {
              "$ref": "#/definitions/ActivityStatsResponse"
            }
          },
          "400": {
            "description": "Failed to parse or request validation failed",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          }
        }
      }
    },
    "/node/provider/consumers-count": {
      "get": {
        "description": "Node unique consumers count served during a period of time.",
        "tags": [
          "provider"
        ],
        "summary": "Provides Node consumers number served during a period of time",
        "operationId": "GetProviderConsumersCount",
        "parameters": [
          {
            "type": "string",
            "description": "period of time (\"1d\", \"7d\", \"30d\")",
            "name": "range",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Provider consumers count",
            "schema": {
              "$ref": "#/definitions/ProviderConsumersCountResponse"
            }
          },
          "400": {
            "description": "Failed to parse or request validation failed",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          }
        }
      }
    },
    "/node/provider/quality": {
      "get": {
        "description": "Node connectivity quality",
        "tags": [
          "provider"
        ],
        "summary": "Provides Node quality",
        "operationId": "GetProviderQuality",
        "responses": {
          "200": {
            "description": "Provider quality",
            "schema": {
              "$ref": "#/definitions/QualityInfoResponse"
            }
          },
          "400": {
            "description": "Failed to parse or request validation failed",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          }
        }
      }
    },
    "/node/provider/series/data": {
      "get": {
        "description": "Node data series metrics of transferred bytes during a period of time.",
        "tags": [
          "provider"
        ],
        "summary": "Provides Node data series metrics of transferred bytes",
        "operationId": "GetProviderTransferredDataSeries",
        "parameters": [
          {
            "type": "string",
            "description": "period of time (\"1d\", \"7d\", \"30d\")",
            "name": "range",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Provider time series metrics of transferred bytes",
            "schema": {
              "$ref": "#/definitions/ProviderTransferredDataSeriesResponse"
            }
          },
          "400": {
            "description": "Failed to parse or request validation failed",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          }
        }
      }
    },
    "/node/provider/series/earnings": {
      "get": {
        "description": "Node time series metrics of earnings during a period of time.",
        "tags": [
          "provider"
        ],
        "summary": "Provides Node  time series metrics of earnings during a period of time",
        "operationId": "GetProviderEarningsSeries",
        "parameters": [
          {
            "type": "string",
            "description": "period of time (\"1d\", \"7d\", \"30d\")",
            "name": "range",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Provider time series metrics of MYSTT earnings",
            "schema": {
              "$ref": "#/definitions/ProviderEarningsSeriesResponse"
            }
          },
          "400": {
            "description": "Failed to parse or request validation failed",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          }
        }
      }
    },
    "/node/provider/series/sessions": {
      "get": {
        "description": "Node time series metrics of sessions started during a period of time.",
        "tags": [
          "provider"
        ],
        "summary": "Provides Node data series metrics of sessions started during a period of time",
        "operationId": "GetProviderSessionsSeries",
        "parameters": [
          {
            "type": "string",
            "description": "period of time (\"1d\", \"7d\", \"30d\")",
            "name": "range",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Provider time series metrics of started sessions",
            "schema": {
              "$ref": "#/definitions/ProviderSessionsSeriesResponse"
            }
          },
          "400": {
            "description": "Failed to parse or request validation failed",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          }
        }
      }
    },
    "/node/provider/service-earnings": {
      "get": {
        "description": "Node earnings per service and total earnings in the all network.",
        "tags": [
          "provider"
        ],
        "summary": "Provides Node earnings per service and total earnings in the all network",
        "operationId": "GetProviderServiceEarnings",
        "responses": {
          "200": {
            "description": "earnings per service and total earnings",
            "schema": {
              "$ref": "#/definitions/EarningsPerServiceResponse"
            }
          },
          "400": {
            "description": "Failed to parse or request validation failed",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          }
        }
      }
    },
    "/node/provider/sessions": {
      "get": {
        "description": "Node sessions metrics during a period of time",
        "tags": [
          "provider"
        ],
        "summary": "Provides Node sessions data during a period of time",
        "operationId": "GetProviderSessions",
        "parameters": [
          {
            "type": "string",
            "description": "period of time (\"1d\", \"7d\", \"30d\")",
            "name": "range",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Provider sessions list",
            "schema": {
              "$ref": "#/definitions/ProviderSessionsResponse"
            }
          },
          "400": {
            "description": "Failed to parse or request validation failed",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          }
        }
      }
    },
    "/node/provider/sessions-count": {
      "get": {
        "description": "Node sessions count during a period of time",
        "tags": [
          "provider"
        ],
        "summary": "Provides Node sessions number during a period of time",
        "operationId": "GetProviderSessionsCount",
        "parameters": [
          {
            "type": "string",
            "description": "period of time (\"1d\", \"7d\", \"30d\")",
            "name": "range",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Provider sessions count",
            "schema": {
              "$ref": "#/definitions/ProviderSessionsCountResponse"
            }
          },
          "400": {
            "description": "Failed to parse or request validation failed",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          }
        }
      }
    },
    "/node/provider/transferred-data": {
      "get": {
        "description": "Node transferred data during a period of time",
        "tags": [
          "provider"
        ],
        "summary": "Provides total traffic served by the provider during a period of time",
        "operationId": "GetProviderTransferredData",
        "parameters": [
          {
            "type": "string",
            "description": "period of time (\"1d\", \"7d\", \"30d\")",
            "name": "range",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Provider transferred data",
            "schema": {
              "$ref": "#/definitions/ProviderTransferredDataResponse"
            }
          },
          "400": {
            "description": "Failed to parse or request validation failed",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/APIError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "APIError": {
      "type": "object",
      "title": "APIError represents an error response from REST API service.",
      "properties": {
        "error": {
          "$ref": "#/definitions/Err"
        },
        "path": {
          "type": "string",
          "x-go-name": "Path"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        }
      },
      "x-go-name": "apiErrorSwagger",
      "x-go-package": "github.com/mysteriumnetwork/go-rest/apierror"
    },
...
    "ActivityStatsResponse": {
      "type": "object",
      "title": "ActivityStatsResponse reflects a node activity stats.",
      "properties": {
        "active_percent": {
          "type": "number",
          "format": "double",
          "x-go-name": "Active"
        },
        "online_percent": {
          "type": "number",
          "format": "double",
          "x-go-name": "Online"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
...
    "LatestReleaseResponse": {
      "description": "LatestReleaseResponse latest release info",
      "type": "object",
      "properties": {
        "version": {
          "type": "string",
          "x-go-name": "Version"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
...
    "ProviderConsumersCountResponse": {
      "type": "object",
      "title": "ProviderConsumersCountResponse reflects a number of unique consumers served during a period of time.",
      "properties": {
        "count": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Count"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
    "ProviderEarningsSeriesResponse": {
      "type": "object",
      "title": "ProviderEarningsSeriesResponse reflects a earnings series metrics during a period of time.",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProviderSeriesItem"
          },
          "x-go-name": "Data"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
...
    "ProviderSessionsResponse": {
      "type": "object",
      "title": "ProviderSessionsResponse reflects a list of sessions metrics during a period of time.",
      "properties": {
        "sessions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProviderSession"
          },
          "x-go-name": "Sessions"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
    "ProviderSessionsSeriesResponse": {
      "type": "object",
      "title": "ProviderSessionsSeriesResponse reflects a sessions data series metrics during a period of time.",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProviderSeriesItem"
          },
          "x-go-name": "Data"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
...
    "ProviderTransferredDataSeriesResponse": {
      "type": "object",
      "title": "ProviderTransferredDataSeriesResponse reflects a transferred bytes data series metrics during a period of time.",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProviderSeriesItem"
          },
          "x-go-name": "Data"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
...
    "QualityInfoResponse": {
      "type": "object",
      "title": "QualityInfoResponse reflects a node quality.",
      "properties": {
        "quality": {
          "type": "number",
          "format": "double",
          "x-go-name": "Quality"
        }
      },
      "x-go-package": "github.com/mysteriumnetwork/node/tequilapi/contract"
    },
...
  }
}